// src/views/availabilityMonitoring/overview/interfaces.ts
import type { ILineModuleGetLineChartRes, ILineModuleGetBarChartRes } from '~/interfaces/common'

/** 图表数据 */
export interface IChartsDataItem {
  overview1: any
  overview2: ILineModuleGetBarChartRes
  overview3: ILineModuleGetBarChartRes
  overview4: ILineModuleGetBarChartRes
  overview5: ILineModuleGetBarChartRes
  overview6: ILineModuleGetLineChartRes
  overview7: ILineModuleGetLineChartRes
  overview8: any
}